updated details page, moved database interactions to background thread
