+++ opensrp-path/src/main/java/org/smartregister/path/tabfragments/ChildRegistrationDataFragment.java
@@ -12,8 +12,6 @@ import org.smartregister.commonregistry.CommonPersonObjectClient;
 import org.smartregister.path.R;
 import org.smartregister.path.activity.ChildDetailTabbedActivity;
 import org.smartregister.path.helper.LocationHelper;
-import org.smartregister.path.viewcomponents.WidgetFactory;
-import org.smartregister.repository.DetailsRepository;
 import org.smartregister.util.DateUtil;
 import org.smartregister.util.Utils;
 import org.smartregister.view.customcontrols.CustomFontTextView;
@@ -62,16 +60,8 @@ public class ChildRegistrationDataFragment extends Fragment {
         return fragmentView;
     }
 
-    @Override
-    public void onResume() {
-        super.onResume();
-
-        loadData();
-    }
-
-    public void loadData() {
+    public void loadData(Map<String, String> detailsMap) {
         if (fragmentView != null) {
-            Map<String, String> detailsMap;
 
             CustomFontTextView tvChildsHomeHealthFacility = (CustomFontTextView) fragmentView.findViewById(R.id.value_childs_home_health_facility);
             CustomFontTextView tvChildsZeirID = (CustomFontTextView) fragmentView.findViewById(R.id.value_childs_zeir_id);
@@ -105,12 +95,7 @@ public class ChildRegistrationDataFragment extends Fragment {
             TableRow tableRowChildsOtherBirthFacility = (TableRow) fragmentView.findViewById(R.id.tableRow_childRegDataFragment_childsOtherBirthFacility);
             TableRow tableRowChildsOtherResidentialArea = (TableRow) fragmentView.findViewById(R.id.tableRow_childRegDataFragment_childsOtherResidentialArea);
 
-            DetailsRepository detailsRepository = ((ChildDetailTabbedActivity) getActivity()).getDetailsRepository();
-            childDetails = childDetails != null ? childDetails : ((ChildDetailTabbedActivity) getActivity()).getChildDetails();
-            detailsMap = detailsRepository.getAllDetailsForClient(childDetails.entityId());
-
             Map<String, String> childDetailsColumnMaps = childDetails.getColumnmaps();
-            WidgetFactory wd = new WidgetFactory();
 
             tvChildsHomeHealthFacility.setText(LocationHelper.getInstance().getOpenMrsReadableName(LocationHelper.getInstance().getOpenMrsLocationName(Utils.getValue(detailsMap, "Home_Facility", false))));
             tvChildsZeirID.setText(Utils.getValue(childDetailsColumnMaps, "zeir_id", false));
